+++ src/configs/Configs.ts
@@ -210,6 +210,7 @@ class Configs {
   static multiSigCleanUpInterval = 120; // seconds
   static tssInstanceRestartGap = 5; // seconds
   static tssUpdateInterval = 10; // seconds
+  static detectionUpdateInterval = 10; // seconds
   static timeoutProcessorInterval = getConfigIntKeyOrDefault(
     'intervals.timeoutProcessorInterval',
     3600
@@ -312,6 +313,12 @@ class Configs {
   static ethCriticalThreshold = BigInt(
     config.get<string>('healthCheck.asset.eth.criticalThreshold')
   );
+  static bnbWarnThreshold = BigInt(
+    config.get<string>('healthCheck.asset.bnb.warnThreshold')
+  );
+  static bnbCriticalThreshold = BigInt(
+    config.get<string>('healthCheck.asset.bnb.criticalThreshold')
+  );
   static ergoScannerWarnDiff = getConfigIntKeyOrDefault(
     'healthCheck.ergoScanner.warnDifference',
     5
@@ -322,11 +329,19 @@ class Configs {
   );
   static ethereumScannerWarnDiff = getConfigIntKeyOrDefault(
     'healthCheck.ethereumScanner.warnDifference',
-    5
+    50
   );
   static ethereumScannerCriticalDiff = getConfigIntKeyOrDefault(
     'healthCheck.ethereumScanner.criticalDifference',
-    20
+    200
+  );
+  static binanceScannerWarnDiff = getConfigIntKeyOrDefault(
+    'healthCheck.binanceScanner.warnDifference',
+    200
+  );
+  static binanceScannerCriticalDiff = getConfigIntKeyOrDefault(
+    'healthCheck.binanceScanner.criticalDifference',
+    800
   );
   static ergoNodeMaxHeightDiff = getConfigIntKeyOrDefault(
     'healthCheck.ergoNode.maxHeightDifference',