+++ src/jobs/runProcessors.ts
@@ -7,6 +7,10 @@ import ColdStorageConfig from '../coldStorage/ColdStorageConfig';
 import TxAgreement from '../agreement/TxAgreement';
 import ArbitraryProcessor from '../arbitrary/ArbitraryProcessor';
 import EventSynchronization from '../synchronization/EventSynchronization';
+import DetectionHandler from '../handlers/DetectionHandler';
+import { DefaultLoggerFactory } from '@rosen-bridge/abstract-logger';
+
+const logger = DefaultLoggerFactory.getInstance().getLogger(import.meta.url);
 
 /**
  * sends generated tx to agreement
@@ -46,6 +50,8 @@ const roundJob = async () => {
     (await TxAgreement.getInstance()).clearTransactions,
     GuardTurn.UP_TIME_LENGTH * 1000
   );
+  // TODO: There are some concerns about sequential execution of Tx Generation jobs
+  //  local:ergo/rosen-bridge/guard-service#317
   // process confirmed events
   await EventProcessor.processConfirmedEvents();
   // process arbitrary orders
@@ -118,6 +124,21 @@ const eventSyncJob = async () => {
 };
 
 /**
+ * runs Detection update job
+ */
+const detectionUpdateJob = () => {
+  DetectionHandler.getInstance()
+    .update()
+    .then(() =>
+      setTimeout(detectionUpdateJob, Configs.detectionUpdateInterval * 1000)
+    )
+    .catch((e) => {
+      logger.error(`Detection update job failed with error: ${e}`);
+      setTimeout(detectionUpdateJob, Configs.detectionUpdateInterval * 1000);
+    });
+};
+
+/**
  * runs all processors and their related jobs
  */
 const runProcessors = () => {
@@ -131,6 +152,7 @@ const runProcessors = () => {
     Configs.requeueWaitingEventsInterval * 1000
   );
   setTimeout(eventSyncJob, Configs.eventSyncInterval * 1000);
+  setTimeout(detectionUpdateJob, Configs.detectionUpdateInterval * 1000);
 };
 
 export { runProcessors };