+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # guard-service
 
+## 5.0.3
+
+### Patch Changes
+
+- Fix commitment verification
+
+## 5.0.2
+
+### Patch Changes
+
+- Validate not merged commitments before create reward transaction
+
 ## 5.0.1
 
 ### Patch Changes